pyproject.toml
[project]
name = "verstor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "agno>=2.5.16",
    "azure-storage-blob>=12.28.0",
    "conf-py==0.1.0",
    "pydantic>=2.13.0",
    "semver>=3.0.4",
    "typer>=0.24.1",
]

[build-system]
requires = ["setuptools>=80.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["verstor*"]
exclude = ["libs*", "tests*", "vendor*"]

[tool.uv.sources]
conf-py = { path = "libs/conf_py-0.1.0-py3-none-any.whl" }

[dependency-groups]
dev = [
    "basedpyright>=1.39.0",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]

[tool.basedpyright]
typeCheckingMode = "strict"
exclude = ["**/versions/*", ".venv", "venv"]

[tool.ruff]
line-length = 120
target-version = "py314"
extend-exclude = ["venv", ".venv", "versions"]
